Extract all lorem stuff into lorem.py
